+++ SplunkAppForWazuh/appserver/static/js/controllers/settings/api/settingsApiCtrl.js
@@ -18,6 +18,7 @@ define(['../../module'], function (controllers) {
       this.scope.currentEntryKey = ''
       this.userRegEx = new RegExp(/^.{3,100}$/)
       this.passRegEx = new RegExp(/^.{3,100}$/)
+      this.aliasRegEx = new RegExp(/^.{3,100}$/)
       this.urlRegEx = new RegExp(/^https?:\/\/[a-zA-Z0-9-.]{1,300}$/)
       this.urlRegExIP = new RegExp(
         /^https?:\/\/[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$/
@@ -27,6 +28,7 @@ define(['../../module'], function (controllers) {
       this.currentDataService = $currentDataService
       this.notification = $notificationService
       this.savingApi = false
+      this.scope.runAs = false;
     }
 
     /**
@@ -38,18 +40,22 @@ define(['../../module'], function (controllers) {
       this.scope.checkManager = entry => this.checkManager(entry)
       this.scope.editEntry = entry => this.editEntry(entry)
       this.scope.removeManager = entry => this.removeManager(entry)
-      this.scope.updateEntry = (user, pass, url, port) =>
-        this.updateEntry(user, pass, url, port)
+      this.scope.updateEntry = (user, pass, url, port, alias, runAs) =>
+        this.updateEntry(user, pass, url, port, alias, runAs)
       this.scope.selectManager = mg => this.selectManager(mg)
-      this.scope.submitApiForm = (user, api, url, port) =>
-        this.submitApiForm(user, api, url, port)
+      this.scope.submitApiForm = (user, api, url, port, alias, runAs) =>
+        this.submitApiForm(user, api, url, port, alias, runAs)
+      this.scope.getIconAndTooltip = entry => this.getIconAndTooltip(entry)
       this.init()
+
+      // Init form values
+      this.clearForm();
 
       // Listens for changes in the selected API
       this.scope.$on('APIChanged', (event, key) => {
         this.setYellowStar(key)
         this.scope.$applyAsync()
-      })      
+      })
     }
 
     /**
@@ -58,10 +64,10 @@ define(['../../module'], function (controllers) {
     async init() {
       try {
         const kvStoreError = ((this.apiList || {}).messages || [])[0] || {}
-        if (kvStoreError.text){
-          this.notification.showErrorToast(kvStoreError.text)
+        if (kvStoreError.text) {
           this.scope.kvStoreInitializing = true
-        }else {
+          throw new Error(kvStoreError.text)
+        } else {
           this.scope.kvStoreInitializing = false
           // If no API, then remove cookie
           if (Array.isArray(this.apiList) && this.apiList.length === 0) {
@@ -89,7 +95,9 @@ define(['../../module'], function (controllers) {
           }
         }
       } catch (err) {
-        this.notification.showErrorToast('Error loading data')
+        this.notification.showErrorToast(
+          `Error loading the data: ${err.message || err}`
+        )
       }
     }
 
@@ -116,7 +124,7 @@ define(['../../module'], function (controllers) {
       } catch (err) {
         this.scope.loadingVizz = false
         this.notification.showErrorToast(
-          `Cannot remove API: ${err.message || err}`
+          `Cannot remove the API: ${err.message || err}`
         )
       }
     }
@@ -134,7 +142,8 @@ define(['../../module'], function (controllers) {
         for (let i = 0; i < this.scope.apiList.length; i++) {
           if (this.scope.apiList[i]._key === entry._key) {
             this.scope.apiList[i] = connectionData
-            this.scope.apiList[i].selected = entry.selected // Check if the API was selected, if it was, set the yellow star
+            // Check if the API was selected, if it was, set the yellow star
+            this.scope.apiList[i].selected = entry.selected
             break
           }
         }
@@ -143,7 +152,9 @@ define(['../../module'], function (controllers) {
         this.scope.$applyAsync()
       } catch (err) {
         this.scope.loadingVizz = false
-        this.notification.showErrorToast(err || 'Unreachable API')
+        this.notification.showErrorToast(
+          `Unreachable API: ${err.message || err}`
+        )
       }
     }
 
@@ -151,6 +162,7 @@ define(['../../module'], function (controllers) {
      * Set form visible
      */
     addNewApiClick() {
+      this.clearForm()
       this.scope.showForm = !this.scope.showForm
       this.scope.edit = false
       this.scope.currentEntryKey = false
@@ -166,16 +178,20 @@ define(['../../module'], function (controllers) {
           this.scope.currentEntryKey === entry['_key'] ? !this.scope.edit : true
         this.scope.showForm = false
         this.scope.currentEntryKey = entry['_key']
+        this.scope.alias = entry.alias
         this.scope.url = entry.url
         this.scope.pass = entry.pass
         this.scope.port = entry.portapi
+        this.scope.runAs = entry.runAs;
         this.scope.user = entry.userapi
         this.scope.managerName = entry.managerName
         this.scope.filterType = entry.filterType
         this.scope.filterName = entry.filterName
         this.scope.entry.url = entry
       } catch (err) {
-        this.notification.showErrorToast('Could not open API form')
+        this.notification.showErrorToast(
+          `Could not open the API form: ${err.message || err}`
+        )
       }
     }
 
@@ -183,13 +199,15 @@ define(['../../module'], function (controllers) {
      * Edits an entry
      * @param {Object} entry
      */
-    async updateEntry(user, pass, url, port) {
+    async updateEntry(user, pass, url, port, alias, runAs = false) {
       try {
         this.scope.loadingVizz = true
         if (this.savingApi) {
           this.notification.showWarningToast('Please, wait for success message')
           return
         }
+        this.scope.validatingError = []
+        this.validateFormInput(user, pass, url, port, alias)
         this.savingApi = true
         this.scope.edit = !this.scope.edit
         this.scope.showForm = false
@@ -197,6 +215,8 @@ define(['../../module'], function (controllers) {
         this.scope.entry.portapi = port
         this.scope.entry.userapi = user
         this.scope.entry.passapi = pass
+        this.scope.entry.alias = alias
+        this.scope.entry.runAs = runAs
         this.scope.entry.filterType = this.scope.filterType
         this.scope.entry.filterName = this.scope.filterName
         this.scope.entry.managerName = this.scope.managerName
@@ -204,7 +224,22 @@ define(['../../module'], function (controllers) {
 
         delete this.scope.entry['$$hashKey']
         await this.currentDataService.checkRawConnection(this.scope.entry)
-        await this.currentDataService.update(this.scope.entry)
+
+        // Update API
+        let res = await this.currentDataService.update(this.scope.entry)
+        // Handle errors
+        if ('data' in res && 'messages' in res.data) {
+          // messages in an array of {type, text} objects
+          res.data.messages.forEach(item => {
+            if ('type' in item && item.type === "ERROR") {
+              throw new Error(item.text)
+            }
+          })
+        }
+
+        // Notify observers (modal)
+        this.scope.$emit('updatedAPI', {})
+
         const updatedApi = await this.currentDataService.checkApiConnection(
           this.scope.entry['_key']
         )
@@ -220,15 +255,18 @@ define(['../../module'], function (controllers) {
           this.currentDataService.getApi() &&
           this.currentDataService.getApi()['_key'] === this.scope.entry['_key']
         ) {
-          this.selectManager(updatedApi['_key'])
+          this.selectManager(updatedApi['_key'], false)
         }
 
         this.scope.edit = false
         this.scope.loadingVizz = false
-        this.notification.showSuccessToast('Updated API')
+        this.notification.showSuccessToast('API updated')
       } catch (err) {
         this.scope.loadingVizz = false
-        this.notification.showErrorToast(err || 'Cannot update API')
+
+        this.notification.showErrorToast(
+          `Cannot update the API: ${err.message || err}`
+        )
       }
       this.savingApi = false
     }
@@ -237,7 +275,7 @@ define(['../../module'], function (controllers) {
      * Select an API as the default one
      * @param {String} key
      */
-    async selectManager(key) {
+    async selectManager(key, showToast = true) {
       try {
         this.scope.loadingVizz = true
         // checking if the api is up
@@ -246,19 +284,43 @@ define(['../../module'], function (controllers) {
         await this.currentDataService.chose(key)
         this.setYellowStar(key)
         this.scope.loadingVizz = false
-        this.notification.showSuccessToast('API selected')
+        if (showToast)
+          this.notification.showSuccessToast('API selected')
         this.scope.$emit('updatedAPI', () => { })
         this.scope.$applyAsync()
       } catch (err) {
         this.scope.loadingVizz = false
-        this.notification.showErrorToast(err || 'Could not select manager')
+        this.notification.showErrorToast(
+          `Could not select manager: ${err.message || err}`
+        )
+      }
+    }
+
+    /**
+     * If values are not valid then throw an error
+     * @param {*} user
+     * @param {*} pass
+     * @param {*} url
+     * @param {*} port
+     * @param {*} alias
+     */
+    validateFormInput(user, pass, url, port, alias) {
+      if (
+        !this.validPassword(pass) ||
+        !this.validPort(port) ||
+        !this.validUrl(url) ||
+        !this.validUsername(user) ||
+        !this.validAlias(alias)
+      ) {
+        this.scope.$applyAsync()
+        throw new Error('Invalid format. Please check the fields again')
       }
     }
 
     /**
      * Adds a new API
      */
-    async submitApiForm(user, pass, url, port) {
+    async submitApiForm(user, pass, url, port, alias, runAs = false) {
       try {
         this.scope.loadingVizz = true
         this.scope.validatingError = []
@@ -268,29 +330,23 @@ define(['../../module'], function (controllers) {
         }
         this.savingApi = true
 
-        // When the Submit button is clicked, get all the form fields by accessing to the input values
+        this.validateFormInput(user, pass, url, port, alias)
+        // When the Submit button is clicked, get all the form
+        // fields by accessing to the input values
         const form_url = url
         const form_apiport = port
         const form_apiuser = user
         const form_apipass = pass
-
-        // If values are not valid then throw an error
-        if (
-          !this.validPassword(form_apipass) ||
-          !this.validPort(form_apiport) ||
-          !this.validUrl(form_url) ||
-          !this.validUsername(form_apiuser)
-        ) {
-          this.scope.$applyAsync()
-          throw new Error('Invalid format. Please check the fields again')
-        }
+        const form_alias = alias
 
         // Create an object to store the field names and values
         const record = {
           url: form_url,
           portapi: form_apiport,
           userapi: form_apiuser,
-          passapi: form_apipass
+          passapi: form_apipass,
+          alias: form_alias,
+          runAs: runAs
         }
 
         // If connected to the API then continue
@@ -304,10 +360,10 @@ define(['../../module'], function (controllers) {
 
         // Check and refresh data
         this.checkManager(api)
-        
+
         // If the only one API in the list, then try to select it
         if (this.scope.apiList.length === 1) {
-          this.selectManager(api['_key'])
+          this.selectManager(api['_key'], false)
         }
 
         this.scope.showForm = false
@@ -325,7 +381,7 @@ define(['../../module'], function (controllers) {
           this.scope.$applyAsync()
         } else {
           this.notification.showErrorToast(
-            err.message || err || 'Cannot save the API.'
+            `Cannot save the API: ${err.message || err}`
           )
         }
       }
@@ -337,10 +393,10 @@ define(['../../module'], function (controllers) {
      * @param {String} url
      */
     validUrl(url) {
-      if (this.urlRegEx.test(url) || this.urlRegExIP.test(url)) {
+      if (url && (this.urlRegEx.test(url) || this.urlRegExIP.test(url))) {
         return true
       } else {
-        this.scope.validatingError.push('Invalid url format')
+        this.scope.validatingError.push('Invalid URL format')
         return false
       }
     }
@@ -350,7 +406,7 @@ define(['../../module'], function (controllers) {
      * @param {String} port
      */
     validPort(port) {
-      if (this.portRegEx.test(port)) {
+      if (port && this.portRegEx.test(port)) {
         return true
       } else {
         this.scope.validatingError.push('Invalid port format')
@@ -363,7 +419,7 @@ define(['../../module'], function (controllers) {
      * @param {String} user
      */
     validUsername(user) {
-      if (this.userRegEx.test(user)) {
+      if (user && this.userRegEx.test(user)) {
         return true
       } else {
         this.scope.validatingError.push(
@@ -372,13 +428,26 @@ define(['../../module'], function (controllers) {
         return false
       }
     }
-
+    /**
+     * Check if an user is valid or not
+     * @param {String} user
+     */
+    validAlias(alias) {
+      if (alias && this.aliasRegEx.test(alias)) {
+        return true
+      } else {
+        this.scope.validatingError.push(
+          'Invalid API alias format, it must have a length between 3 and 100 characters.'
+        )
+        return false
+      }
+    }
     /**
      * Check if a password is valid or not
      * @param {String} pass
      */
     validPassword(pass) {
-      if (this.passRegEx.test(pass)) {
+      if (pass && this.passRegEx.test(pass)) {
         return true
       } else {
         this.scope.validatingError.push(
@@ -396,6 +465,8 @@ define(['../../module'], function (controllers) {
       this.scope.port = ''
       this.scope.user = ''
       this.scope.pass = ''
+      this.scope.alias = ''
+      this.scope.runAs = false
     }
 
     setYellowStar(key) {
@@ -403,6 +474,29 @@ define(['../../module'], function (controllers) {
         api['_key'] === key ? (api.selected = true) : (api.selected = false)
       })
     }
+
+
+    /**
+     * Evaluates the class and tooltip text for the given API 
+     * object depending on the `run_as` value.
+     * 
+     * If the `run_as` field is true, a check icon is shown. If it is false,
+     * a cross icon is shown instead. Respectively, the tooltip text takes 
+     * 'enabled' or 'disable'.
+     * 
+     * @param {Object} entry API object to evaluate.
+     * @returns Object with the CSS class for the icon and string text 
+     * for the tooltip.
+     */
+    getIconAndTooltip(entry) {
+      // Cast to boolean
+      const runAs = (entry.runAs === "true") || (entry.runAs === true)
+      return {
+        class: runAs ? 'fa fa-check' : 'fa fa-times',
+        tooltip: runAs ? 'enabled' : 'disabled'
+      }
+    }
+
   }
   controllers.controller('settingsApiCtrl', SettingsApi)
 })